refactor(fabric): isolate 26.1 platform code under mc261
Move the Fabric 26.1 grimMainLoad slot from package mc1216 to mc261 with
clear Fabric261* type names and GrimACFabric261LoaderPlugin. Wire
fabric.mod.json to the new entrypoint and report ServerVersion.V_26_1.

Remove the duplicate fabric/mc12111 Gradle tree (unused in settings.gradle)
to avoid two sources of truth.

Made-with: Cursor
