--- fabric/mc12111/src/main/java/ac/grim/grimac/platform/fabric/mc1216/GrimACFabric1212LoaderPlugin.java
@@ -1,45 +1,0 @@
-package ac.grim.grimac.platform.fabric.mc1216;
-
-import ac.grim.grimac.platform.fabric.mc1216.command.Fabric1212PlayerSelectorAdapter;
-import ac.grim.grimac.platform.fabric.command.FabricPlayerSelectorParser;
-import ac.grim.grimac.platform.fabric.manager.FabricParserDescriptorFactory;
-import ac.grim.grimac.platform.fabric.mc1194.GrimACFabric1190LoaderPlugin;
-import ac.grim.grimac.platform.fabric.mc1194.entity.Fabric1194GrimEntity;
-import ac.grim.grimac.platform.fabric.mc1194.player.Fabric1193PlatformInventory;
-import ac.grim.grimac.platform.fabric.mc1205.Fabric1203PlatformServer;
-import ac.grim.grimac.platform.fabric.mc1205.convert.Fabric1200MessageUtil;
-import ac.grim.grimac.platform.fabric.mc1205.convert.Fabric1205ConversionUtil;
-import ac.grim.grimac.platform.fabric.mc1216.convert.Fabric1216ConversionUtil;
-import ac.grim.grimac.platform.fabric.mc1216.player.Fabric1212PlatformPlayer;
-import ac.grim.grimac.platform.fabric.mc1216.player.Fabric1215PlatformInventory;
-import ac.grim.grimac.platform.fabric.player.FabricPlatformPlayerFactory;
-import ac.grim.grimac.utils.lazy.LazyHolder;
-import com.github.retrooper.packetevents.PacketEvents;
-import com.github.retrooper.packetevents.manager.server.ServerVersion;
-
-public class GrimACFabric1212LoaderPlugin extends GrimACFabric1190LoaderPlugin {
-
-    public GrimACFabric1212LoaderPlugin() {
-        super(
-                LazyHolder.simple(() -> new FabricParserDescriptorFactory(
-                        new FabricPlayerSelectorParser<>(Fabric1212PlayerSelectorAdapter::new)
-                )),
-                new FabricPlatformPlayerFactory(
-                        Fabric1212PlatformPlayer::new,
-                        Fabric1194GrimEntity::new,
-                        PacketEvents.getAPI().getServerManager().getVersion().isNewerThan(ServerVersion.V_1_21_4)
-                            ? Fabric1215PlatformInventory::new : Fabric1193PlatformInventory::new
-                ),
-                PacketEvents.getAPI().getServerManager().getVersion().isNewerThan(ServerVersion.V_1_21_10) ?
-                        new Fabric12111PlatformServer() : new Fabric1203PlatformServer(),
-                new Fabric1200MessageUtil(),
-                PacketEvents.getAPI().getServerManager().getVersion().isNewerThan(ServerVersion.V_1_21_5)
-                        ? new Fabric1216ConversionUtil() : new Fabric1205ConversionUtil()
-        );
-    }
-
-    @Override
-    public ServerVersion getNativeVersion() {
-        return ServerVersion.V_1_21_11;
-    }
-}
