--- fabric/mc12111/src/main/resources/fabric.mod.json
@@ -1,22 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "grimac-mc1215",
-  "version": "${version}",
-  "name": "GrimAC for 1.21.2+ built against 1.21.11",
-  "description": "Libre simulation anticheat designed for 1.21 with 1.8-1.21 support, powered by PacketEvents 2.0.",
-  "authors": [
-    "GrimAC"
-  ],
-  "license": "GPLv3",
-  "environment": "server",
-  "entrypoints": {
-    "grimMainLoad": [
-      "ac.grim.grimac.platform.fabric.mc1216.GrimACFabric1212LoaderPlugin"
-    ]
-  },
-  "accessWidener": "grimac.accesswidener",
-  "depends": {
-    "minecraft": ">=1.21.2",
-    "fabric-lifecycle-events-v1": "*"
-  }
-}
