--- fabric/src/main/java/ac/grim/grimac/platform/fabric/mc1216/Fabric12111PlatformServer.java
@@ -1,10 +1,0 @@
-package ac.grim.grimac.platform.fabric.mc1216;
-
-import ac.grim.grimac.platform.fabric.mc1205.Fabric1203PlatformServer;
-
-/**
- * Intentional marker subclass for 1.21.11 platform selection.
- * No additional behavior is needed yet; it delegates to Fabric1203PlatformServer.
- */
-public class Fabric12111PlatformServer extends Fabric1203PlatformServer {
-}
