--- fabric/src/main/java/ac/grim/grimac/platform/fabric/mc1216/player/Fabric1215PlatformInventory.java
@@ -1,10 +1,0 @@
-package ac.grim.grimac.platform.fabric.mc1216.player;
-
-import ac.grim.grimac.platform.fabric.mc1194.player.Fabric1193PlatformInventory;
-import ac.grim.grimac.platform.fabric.player.AbstractFabricPlatformPlayer;
-
-public class Fabric1215PlatformInventory extends Fabric1193PlatformInventory {
-    public Fabric1215PlatformInventory(AbstractFabricPlatformPlayer player) {
-        super(player);
-    }
-}
