+++ fabric/src/main/java/ac/grim/grimac/platform/fabric/mc261/Fabric261PlatformServer.java
@@ -1,0 +1,9 @@
+package ac.grim.grimac.platform.fabric.mc261;
+
+import ac.grim.grimac.platform.fabric.mc1205.Fabric1203PlatformServer;
+
+/**
+ * Fabric platform server for Minecraft 26.1+ (official mappings). Extends the 1.20.5-era base;
+ * used by {@link GrimACFabric261LoaderPlugin} when PacketEvents reports a newer protocol than 1.21.4.
+ */
+public class Fabric261PlatformServer extends Fabric1203PlatformServer {}
