+++ fabric/src/main/java/ac/grim/grimac/platform/fabric/mc261/command/Fabric261PlayerSelectorAdapter.java
@@ -1,4 +1,4 @@
-package ac.grim.grimac.platform.fabric.mc1216.command;
+package ac.grim.grimac.platform.fabric.mc261.command;
 
 import ac.grim.grimac.GrimAPI;
 import ac.grim.grimac.platform.api.sender.Sender;
@@ -6,13 +6,13 @@ import ac.grim.grimac.platform.fabric.mc1161.command.Fabric1161PlayerSelectorAda
 import ac.grim.grimac.platform.fabric.sender.FabricSenderFactory;
 import org.incendo.cloud.minecraft.modded.data.SinglePlayerSelector;
 
-public class Fabric1212PlayerSelectorAdapter extends Fabric1161PlayerSelectorAdapter {
+public class Fabric261PlayerSelectorAdapter extends Fabric1161PlayerSelectorAdapter {
 
-    public Fabric1212PlayerSelectorAdapter(SinglePlayerSelector fabricSelector) {
+    public Fabric261PlayerSelectorAdapter(SinglePlayerSelector fabricSelector) {
         super(fabricSelector);
     }
 
-    // 1.21.2 .getCommandSource() moves from entity to player
+    // 1.21.2+: .getCommandSource() moves from entity to player
     @Override
     public Sender getSinglePlayer() {
         return ((FabricSenderFactory) GrimAPI.INSTANCE.getSenderFactory()).wrap(fabricSelector.single().createCommandSourceStack());
