+++ fabric/src/main/java/ac/grim/grimac/platform/fabric/mc261/player/Fabric261PlatformInventory.java
@@ -1,0 +1,10 @@
+package ac.grim.grimac.platform.fabric.mc261.player;
+
+import ac.grim.grimac.platform.fabric.mc1194.player.Fabric1193PlatformInventory;
+import ac.grim.grimac.platform.fabric.player.AbstractFabricPlatformPlayer;
+
+public class Fabric261PlatformInventory extends Fabric1193PlatformInventory {
+    public Fabric261PlatformInventory(AbstractFabricPlatformPlayer player) {
+        super(player);
+    }
+}
